🦀 Comprehensive Clippy run (#380)
## What?
Run clippy::all, which contains the default lint groups (correctness, suspicious, style, complexity, perf)

## Why?
We haven't checked clippy in a long time, it provides useful feedback on code style

## How?
We define in the workspace the clippy lints we want to run, and the ones we want to ignore

In certain pallets like democracy, which are forks, we don't want to check the clippy there, so we ignore everything.

## Anything else?
I started trying to do both all and pedantic groups, but pedantic is huge and would have made the PR bloated.

I implemented some of those changes like the module name repetition, but then stopped.

I will in a later PR cherry pick some lints from pedantic and add them to the workspace