AssetId Storage Migration (#434)
## What?
Migrate all storage that mentioned u32 asset id to a Location

## Why?
Previous PR changed this

## How?
Its technically a new storage item since the asset id was used as a key, so we delete the old storage and create a new one for each item.

## Testing?
we did a manual override with chopsticks and a try-runtime run and it worked.